Refractor Post and add test for coverage
