+++ views/blogs/components/post/Content.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components/macro'
+
+export default styled.div`
+  margin-top: 2em;
+  font-size: 22px;
+  letter-spacing: -0.004em;
+  line-height: 38px;
+  color: rgba(0, 0, 0, 0.84);
+  font-family: Georgia, Cambria, 'Times New Roman', Times, serif;
+`
