+++ views/blogs/components/post/Header.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components/macro'
+
+export default styled.h1`
+  line-height: 48px;
+  font-size: 40px;
+  color: rgba(0, 0, 0, 0.84);
+  font-weight: 400;
+`
